Seperated out the pg_hasher code
